+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 All notable changes to this project will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## [0.9.4] - 2015-05-09
+### Added
+ - New benchmarking information
+ - Plugins panel
+ - Routing panel
+
+### Changed
+ - Refactor and improvements
+
+
 ## [0.9.3] - 2015-05-08
 ### Changed
  - Refactor and improvements